# Persist JIT cache to make tests run faster