Store JIT cache files in the specified directory to make tests run faster.

Without JIT cache:

```
$ time ./run_test.py --test chainer-py2 --cache /tmp/my-cache
1.65s user 3.58s system 1% cpu 6:52.21 total
```

With JIT cache (once JIT cache is generated):

```
$ time ./run_test.py --test chainer-py2 --cache /tmp/my-cache
1.73s user 2.92s system 1% cpu 4:43.07 total
```

Tested under chainer==3.0.0rc1 & cupy==2.0.0rc1.

c.f. https://devblogs.nvidia.com/parallelforall/cuda-pro-tip-understand-fat-binaries-jit-caching/